Commit for the sake of commit.
Abandoned the stackpacnel of richtextbox'es as that was causing OOM. Probably something that can be revisited.

Switched to less satisfying IE control, which as it turns out is opaque and won't show our nice background. Since IE10 for WP8 doesn't support fixed positioning, can't even "fake" the background. Much disappoint.